build: fix double linkage of Seastar libraries in tests
Tests that use the seastar test framework link with its libraries (as described
pkg-config --libs --static seastar-testing.pc). The output of that command however
also includes seastar's libraries, because seastar-testing.pc explicitly depends
on seastar. As a result, the linker command line specifies each seastar library
(and its dependencies) twice.

Up to now this has not been a problem; the linker ignores the doubly-specified
libraries. However with a recent Seastar fix, this starts failing with --enable-dpdk,
because dpdk libraries are linked with --whole-archive. This forces the linker to
include the library in the executable, and if this is done twice, the linker
complains about duplicate symbols.

Fix by removing the --static flag from the pkg-config command line. Since we
explicitly include seastar dependencies, it is not needed. With this we can build
Scylla tests with recent Seastar and with --enable-dpdk.