misc_services: Don't gossip hit rates when they are not used
When cache_hit_rate_read_balancing is set to false,
heat-weighted load balancing is turned off and we
gossip cache hit rates for no good reason.
The size of gossiped data is linear on number of tables
in the database so it's good to avoid such traffic
when it is not used.

Fixes #5183

Signed-off-by: Piotr Jastrzebski <piotr@scylladb.com>